+++ src/com/dhbw/morse_project/gui/Morse_Symbolbild.png
